QResultStore::moveResult(): refuse to move from lvalues
The T&& result argument is deduced, therefore result is a Universal
Reference and as such can bind to lvalues, not just rvalues. If passed
an lvalue, the function would happily move away from it, which is not
what the doctor prescribed.

Catch the attempt with a static assertion.

Change-Id: Iff8c1abd5dcb1043ed94ba76570be5ba3d6f92da
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 10e106ae47c5a9450623d9874fdfd0c0aee54f26)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>